+++ DynamicIsland/models/Constants.swift
@@ -1152,6 +1152,7 @@ extension Defaults.Keys {
     
     // MARK: Timer Feature
     static let enableTimerFeature = Key<Bool>("enableTimerFeature", default: true)
+    static let togglEnabled = Key<Bool>("togglEnabled", default: false)
     static let timerDisplayMode = Key<TimerDisplayMode>("timerDisplayMode", default: .tab)
     static let timerPresets = Key<[TimerPreset]>("timerPresets", default: TimerPreset.defaultPresets)
     static let showTimerPresetsInNotchTab = Key<Bool>("showTimerPresetsInNotchTab", default: true)